Release v1.7.1
Use a regex to match ` + ` in shortcuts.
Resolve #66.
Tweak quickey-ctrl-tab.ahk to exclude non-Chromium browsers and clean up formatting.
Ignore whether we previously saw a ctrl-tab when handling ctrlUp, so that if the menu is already open, pressing ctrl-tab and releasing will select the item.
Update version to 1.7.1.
Update releases/index.md.