Release v2.0.1
Wait for open popup command promises to settle before handling the next one.  If the keyboard shortcut was hit quickly, it would cause the popup to open and then close multiple times.
Fix a timing issue with combining navigate to previous and toggle commands
Make sure calls to navigate to the previous/next tab follow any pending toggles, and vice versa.
Reset the toolbar icon if the toggle command is executed during the navigation cooldown.
Update releases/index.md with changes.
Enable a caller to wait for the next execution of a debounced function and trigger it immediately.
Add toolbar.setNormalIcon() method.
Wrap try/catch around calls to chrome APIs in popup-window.js.
Update babel target to 115, the minimum version in the manifest.
Fix #123.